# src/main/resources/openapi.yaml
openapi: "3.0.2"
info:
  title: API Ajmanre
  version: "1.0"
servers:
  - url: http://localhost:{port}
    description: The local API server
    variables:
      port:
        enum:
          - '8443'
          - '443'
        default: '8443'
  - url: https://angular.ajmanre.com:{port}
    description: The local API server
    variables:
      port:
        enum:
          - '8443'
          - '443'
        default: '8443'
...
paths:
...
  /requirement:
    post:
      tags:
        - "requirementAdd"
      summary: "add requirement"
      operationId: "requirementPost"
      parameters:
        - in: header
          name: Authorization
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Requirement"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /requirement/{page}/{size}:
    get:
      tags:
        - "requirementPageSize"
      summary: "requirement listing"
      operationId: "requirementPageSize"
      parameters:
        - in: path
          name : page
          schema:
            type: integer
            default: 1
          required: true
          description: "page no of the list pagination"
        - in: path
          name : size
          schema:
            type: integer
            default: 5
          required: true
          description: "size of the pagination"
      responses:
        "200":
          description: "requirement pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequirementPage"
...
  /requirement/{id}:
    put:
      tags:
        - "requirementEdit"
      summary: "edit requirement"
      operationId: "requirementPut"
      parameters:
        - in: path
          name : id
          schema:
            type: string
          required: true
          description: "update requirement details by id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Requirement"
      responses:
        "200":
          description: "MessageResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /user/{page}/{size}:
    get:
      tags:
        - "userPageSize"
      summary: "users listing"
      operationId: "userPageSize"
      parameters:
        - in: path
          name : page
          schema:
            type: integer
            default: 1
          required: true
          description: "page no of the list pagination"
        - in: path
          name : size
          schema:
            type: integer
            default: 5
          required: true
          description: "size of the pagination"
      responses:
        "200":
          description: "users pagination listing"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPage"
...
components:
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: "object"
      required:
        - message
      properties:
        message:
          type: "string"
          description: "Message"
        identifier:
          type: "string"
          description: "If identifier has to be in response"
...
    Source:
      type: "object"
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: source identifier
        name:
          type: string
          description: key value of the  source
...
    Requirement:
      type: "object"
      required:
        - title
      properties:
        id:
          type: string
          description: source identifier
        updatedAt:
          type: "string"
          format: "date-time"
        name:
          type: string
          description: title summary
        email:
          type: string
        phoneNo:
          type: string
        whatsapp:
          type: string
          description: whatsapp
        message:
          type: string
          description: application
        featured:
          type: boolean
          description: is feature listing
        listingExpDate:
          type: "string"
          format: "date-time"
        postedDate:
          type: "string"
          format: "date-time"
        postedBy:
          $ref: "#/components/schemas/Source"
    RequirementPage:
      type: "object"
      required:
        - page
        - size
        - total
      properties:
        page:
          type: integer
          description: page no of data result in pagination
        size:
          type: integer
          description: no of data elements in page
        data:
          type: array
          items:
            $ref: "#/components/schemas/Requirement"
          description: data elements list in page
        total:
          type: integer
          format: int64
          description: total elements in the list
...
    UserPage:
      type: "object"
      required:
        - page
        - size
        - total
      properties:
        page:
          type: integer
          description: page no of data result in pagination
        size:
          type: integer
          description: no of data elements in page
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
          description: data elements list in page
        total:
          type: integer
          format: int64
          description: total elements in the list
    User:
      type: "object"
      required:
        - id
        - updatedAt
        - username
        - password
        - email
        - name
      properties:
        id:
          type: "string"
        updatedAt:
          type: "string"
          format: "date-time"
        username:
          type: "string"
          description: "Username"
        password:
          type: "string"
          description: "Password"
        name:
          type: "string"
          description: "name"
        email:
          type: "string"
          description: "email"
        roles:
          type: "array"
          description: "roles, admin, agent, agency, user (individual), owner"
          items:
            type: string
          example: ['admin', 'agent', 'agency', 'owner', 'user']
...
security:
  - bearerAuth: []  